Renamed incorreclty name stest files, updated format of CardItems.test.js, added snapshot tests for CollaboratorsLink.test.js
